DOC: all componentts v.1.2
