+++ pages/_document.js
@@ -1,10 +1,7 @@
 // meta 태그를 정의, 전체페이지의 구조를 만들어준다. (html의 meta)
 // _document작성시 랜더함수는 꼭 <Html>, <Head>, <Main>, <NextScript>요소를 리턴해줘야한다.
 
-import React from 'react';
 import { Html, Head, Main, NextScript } from 'next/document';
-
-const key = process.env.EXAMPLE_KEY; // 환경변수 키 사용법
 
 const Document = () => {
   return (
