+++ pages/mainPages/Home.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Home = () => {
   return <div>홈</div>;
 };
