--- pages/subPages/MyReservation.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const MyReservation = () => {
-  return <div></div>;
-};
-
-export default MyReservation;
