--- src/components/molecules/Banners.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Banners = () => {
-  return <div></div>;
-};
-
-export default Banners;
