--- src/components/molecules/Calender.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Calender = () => {
-  return <div></div>;
-};
-
-export default Calender;
