+++ src/components/molecules/Layout.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import Navbar from './Navbar';
-import { RecoilRoot } from 'recoil';
-import { logout } from '@/pages/api/auth/firebase';
-import { auth } from '../../../pages/api/auth/firebase';
 import { useRouter } from 'next/router';
 import toast, { Toaster } from 'react-hot-toast';
+import { RecoilRoot } from 'recoil';
+import Navbar from './Navbar';
+import { logout } from '../../../pages/api/auth/firebase';
+import { auth } from '../../../pages/api/auth/firebase';
 
 const signOutNotify = () =>
   toast('반가웠어요 잘가요!', {
