--- src/components/template/TemPlate.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const TemPlate = () => {
-  return <div></div>;
-};
-
-export default TemPlate;
