--- src/serverRouter/descriptionRouter.js
@@ -1,25 +1,0 @@
-// const express = require("express");
-// const router = express.Router();
-// const request = require("request");
-// const converter = require("xml-js");
-
-// // 공공기관 api 경로 (쿼리전까지)
-
-// // api key
-// const KEY = "98e02b76a394447699b7324b7ff14b83";
-
-// router.get("/description/:id", (req, res) => {
-//     // request(
-//     //     {
-//     //         url: `http://kopis.or.kr/openApi/restful/pblprfr/${req.params.id}?service=${KEY}`,
-//     //         method: "GET",
-//     //     },
-//     //     (err, response, body) => {
-//     //         const xmlToJson = converter.xml2json(body);
-//     //         res.send(xmlToJson);
-//     //     }
-//     // );
-//     console.log(req.params.id);
-// });
-
-// module.exports = router;
