+++ src/store/server/useGetPerformance.js
@@ -32,7 +32,7 @@ const useGetPerformance = () => {
           data: arr,
         };
       },
-    },
+    }
   );
 };
